# docs/v1/swagger.yaml
basePath: /v1
definitions:
  config.WorkflowModelParam:
    properties:
      default_val:
        $ref: '#/definitions/config.WorkflowModelParamDefaultVal'
      desc:
        type: string
      label:
        type: string
      max:
        type: string
      min:
        type: string
      name:
        type: string
      param_class:
        $ref: '#/definitions/config.WorkflowModelParamClass'
      precision:
        type: integer
      type:
        type: integer
    type: object
  config.WorkflowModelParamClass:
    properties:
      class_id:
        type: integer
      label:
        type: string
    type: object
  config.WorkflowModelParamDefaultVal:
    properties:
      balance:
        type: string
      creative:
        type: string
      default_val:
        type: string
      precise:
        type: string
    type: object
...
  mp.ProviderModelByHuoshan:
    properties:
      embedding:
        $ref: '#/definitions/mp_huoshan.Embedding'
      llm:
        $ref: '#/definitions/mp_huoshan.LLM'
    type: object
  mp.ProviderModelByInfini:
    properties:
      embedding:
        $ref: '#/definitions/mp_infini.Embedding'
      llm:
        $ref: '#/definitions/mp_infini.LLM'
      rerank:
        $ref: '#/definitions/mp_infini.Rerank'
    type: object
  mp.ProviderModelByOllama:
    properties:
      embedding:
        $ref: '#/definitions/mp_ollama.Embedding'
      llm:
        $ref: '#/definitions/mp_ollama.LLM'
    type: object
  mp.ProviderModelByOpenAICompatible:
    properties:
      embedding:
        $ref: '#/definitions/mp_openai_compatible.Embedding'
      llm:
        $ref: '#/definitions/mp_openai_compatible.LLM'
      rerank:
        $ref: '#/definitions/mp_openai_compatible.Rerank'
    type: object
  mp.ProviderModelByQwen:
    properties:
      embedding:
        $ref: '#/definitions/mp_qwen.Embedding'
      llm:
        $ref: '#/definitions/mp_qwen.LLM'
      rerank:
        $ref: '#/definitions/mp_qwen.Rerank'
    type: object
  mp.ProviderModelByYuanjing:
    properties:
      embedding:
        $ref: '#/definitions/mp_yuanjing.Embedding'
      gui:
        $ref: '#/definitions/mp_yuanjing.Gui'
      llm:
        $ref: '#/definitions/mp_yuanjing.LLM'
      ocr:
        $ref: '#/definitions/mp_yuanjing.Ocr'
      pdf-parser:
        $ref: '#/definitions/mp_yuanjing.PdfParser'
      rerank:
        $ref: '#/definitions/mp_yuanjing.Rerank'
    type: object
  mp.ProviderModelConfig:
    properties:
      providerHuoshan:
        $ref: '#/definitions/mp.ProviderModelByHuoshan'
      providerModelByInfini:
        $ref: '#/definitions/mp.ProviderModelByInfini'
      providerOllama:
        $ref: '#/definitions/mp.ProviderModelByOllama'
      providerOpenAICompatible:
        $ref: '#/definitions/mp.ProviderModelByOpenAICompatible'
      providerQwen:
        $ref: '#/definitions/mp.ProviderModelByQwen'
      providerYuanJing:
        $ref: '#/definitions/mp.ProviderModelByYuanjing'
    type: object
  mp_common.Tag:
    properties:
      text:
        type: string
    type: object
  mp_huoshan.Embedding:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_huoshan.LLM:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
      functionCalling:
        description: 函数调用是否支持
        enum:
        - noSupport
        - toolCall
        type: string
      maxTokens:
        description: 模型回答最大tokens
        type: integer
      visionSupport:
        description: 视觉支持
        enum:
        - noSupport
        - support
        type: string
    type: object
...
  mp_infini.Embedding:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_infini.LLM:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
      functionCalling:
        description: 函数调用是否支持
        enum:
        - noSupport
        - toolCall
        type: string
      maxTokens:
        description: 模型回答最大tokens
        type: integer
      visionSupport:
        description: 视觉支持
        enum:
        - noSupport
        - support
        type: string
    type: object
...
  mp_infini.Rerank:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_ollama.Embedding:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_ollama.LLM:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
      functionCalling:
        description: 函数调用是否支持
        enum:
        - noSupport
        - toolCall
        type: string
      maxTokens:
        description: 模型回答最大tokens
        type: integer
      visionSupport:
        description: 视觉支持
        enum:
        - noSupport
        - support
        type: string
    type: object
...
  mp_openai_compatible.Embedding:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_openai_compatible.LLM:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
      functionCalling:
        description: 函数调用是否支持
        enum:
        - noSupport
        - toolCall
        type: string
      maxTokens:
        description: 模型回答最大tokens
        type: integer
      visionSupport:
        description: 视觉支持
        enum:
        - noSupport
        - support
        type: string
    type: object
...
  mp_openai_compatible.Rerank:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_qwen.Embedding:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_qwen.LLM:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
      functionCalling:
        description: 函数调用是否支持
        enum:
        - noSupport
        - toolCall
        type: string
      maxTokens:
        description: 模型回答最大tokens
        type: integer
      visionSupport:
        description: 视觉支持
        enum:
        - noSupport
        - support
        type: string
    type: object
...
  mp_qwen.Rerank:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_yuanjing.Embedding:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_yuanjing.Gui:
    properties:
      apiKey:
        description: ApiKey
        type: string
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_yuanjing.LLM:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
      functionCalling:
        description: 函数调用是否支持
        enum:
        - noSupport
        - toolCall
        type: string
      maxTokens:
        description: 模型回答最大tokens
        type: integer
      visionSupport:
        description: 视觉支持
        enum:
        - noSupport
        - support
        type: string
    type: object
...
  mp_yuanjing.Ocr:
    properties:
      apiKey:
        description: ApiKey
        type: string
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_yuanjing.PdfParser:
    properties:
      apiKey:
        description: ApiKey
        type: string
      endpointUrl:
        description: 推理url
        type: string
    type: object
  mp_yuanjing.Rerank:
    properties:
      apiKey:
        description: ApiKey
        type: string
      contextSize:
        description: 上下文长度
        type: integer
      endpointUrl:
        description: 推理url
        type: string
    type: object
...
  request.AppBriefConfig:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    required:
    - name
    type: object
...
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.CreateWorkflowByTemplateReq:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      templateId:
        type: string
    required:
    - name
    - templateId
    type: object
...
  request.WorkflowIDReq:
    properties:
      workflow_id:
        type: string
    required:
    - workflow_id
    type: object
...
  response.CozeWorkflowIDData:
    properties:
      workflow_id:
        type: string
    type: object
  response.CozeWorkflowModelInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 模型图标路径
      config: {}
      createdAt:
        type: string
      displayName:
        description: 模型显示名称
        type: string
      examples:
        allOf:
        - $ref: '#/definitions/mp.ProviderModelConfig'
        description: 仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数
      isActive:
        description: '启用状态（true: 启用，false: 禁用）'
        type: boolean
      model:
        description: 模型名称
        type: string
      model_ability:
        $ref: '#/definitions/response.CozeWorkflowModelInfoAbility'
      model_params:
        items:
          $ref: '#/definitions/config.WorkflowModelParam'
        type: array
      modelDesc:
        type: string
      modelId:
        type: string
      modelType:
        enum:
        - llm
        - embedding
        - rerank
        type: string
      orgId:
        type: string
      provider:
        description: 模型供应商
        enum:
        - OpenAI-API-compatible
        - YuanJing
        type: string
      publishDate:
        description: 模型发布时间
        type: string
      tags:
        items:
          $ref: '#/definitions/mp_common.Tag'
        type: array
      updatedAt:
        type: string
      userId:
        type: string
    required:
    - model
    - modelType
    - provider
    type: object
  response.CozeWorkflowModelInfoAbility:
    properties:
      audio_understanding:
        type: boolean
      cot_display:
        type: boolean
      function_call:
        type: boolean
      image_understanding:
        type: boolean
      video_understanding:
        type: boolean
    type: object
...
  response.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.ToolAction4Workflow:
    properties:
      actionId:
        type: string
      actionName:
        type: string
      desc:
        type: string
    type: object
...
  response.ToolDetail4Workflow:
    properties:
      actionId:
        type: string
      actionName:
        type: string
      iconUrl:
        type: string
      inputs:
        items: {}
        type: array
      outputs:
        items: {}
        type: array
    type: object
...
  response.ToolSelect4Workflow:
    properties:
      actions:
        items:
          $ref: '#/definitions/response.ToolAction4Workflow'
        type: array
      apiKey:
        type: string
      desc:
        type: string
      iconUrl:
        type: string
      toolId:
        type: string
      toolName:
        type: string
      toolType:
        type: string
    type: object
...
info:
  contact: {}
  description: "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- |
    --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n|
    X-Org-Id      | 组织ID    |\r\n| X-Client-Id   | 客户端标识|\r\n\r\n## HTTP Status\r\n|
    HTTP Status             | 说明                   |\r\n| -----------------------
    | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n|
    400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败
    \           |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n|
    一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n|
    guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      |
    【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n|
    permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      |
    权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n##
    `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n
    \   \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\":
    \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n
    \       {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n
...
    \"操作成功\"\r\n}\r\n```"
...
  version: v0.0.1
paths:
...
  /appspace/workflow:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Workflow的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AppBriefConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CozeWorkflowIDData'
              type: object
      security:
      - JWT: []
      summary: 创建Workflow
      tags:
      - workflow
  /appspace/workflow/copy:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Workflow的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.WorkflowIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CozeWorkflowIDData'
              type: object
      security:
      - JWT: []
      summary: 拷贝Workflow
      tags:
      - workflow
  /appspace/workflow/export:
    get:
      consumes:
      - application/json
      description: 导出工作流的json文件
      parameters:
      - description: 工作流ID
        in: query
        name: workflow_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 导出Workflow
      tags:
      - workflow
  /appspace/workflow/import:
    post:
      consumes:
      - multipart/form-data
      description: 通过JSON文件导入工作流
      parameters:
      - description: 工作流JSON文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CozeWorkflowIDData'
              type: object
      security:
      - JWT: []
      summary: 导入Workflow
      tags:
      - workflow
  /appspace/workflow/model/select/llm:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        $ref: '#/definitions/response.CozeWorkflowModelInfo'
                    type: object
              type: object
      security:
      - JWT: []
      summary: llm模型列表（用于workflow）
      tags:
      - workflow
...
  /workflow/template:
    post:
      consumes:
      - application/json
      parameters:
      - description: 通过模板创建Workflow的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateWorkflowByTemplateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CozeWorkflowIDData'
              type: object
      security:
      - JWT: []
      summary: 复制工作流模板
      tags:
      - workflow
...
  /workflow/tool/action:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工具类型
        enum:
        - builtin
        - custom
        in: query
        name: toolType
        required: true
        type: string
      - description: 工具具体action名称
        in: query
        name: actionName
        required: true
        type: string
      - description: 工具ID
        in: query
        name: toolId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolDetail4Workflow'
              type: object
      security:
      - JWT: []
      summary: 工具具体action（用于workflow）
      tags:
      - workflow
  /workflow/tool/select:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工具类型
        enum:
        - builtin
        - custom
        in: query
        name: toolType
        required: true
        type: string
      - description: 工具名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ToolSelect4Workflow'
                        type: array
                    type: object
              type: object
      security:
      - JWT: []
      summary: 工具列表（用于workflow）
      tags:
      - workflow
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"